+++ client/src/Components/Gameboard/ClueView.jsx
@@ -1,8 +1,8 @@
 import {useState, useContext, useEffect} from 'react';
-import ClueInput from '../ClueInput.jsx';
+import ClueInput from './ClueInput.jsx';
 
 // editor should be true when isSpyMaster and teamTurn matches own team
-const ClueView = ({ shareClue, activeTeam, passTurn, editor = false }) => {
+const ClueView = ({ room, currentTeam, editor = false }) => {
   const [editing, setEditing] = useState(editor);
   const [clue, setClue] = useState('waiting for clue...');
   const [clueNumber, setClueNumber] = useState('');
@@ -11,17 +11,14 @@ const ClueView = ({ shareClue, activeTeam, passTurn, editor = false }) => {
   // This gets called within the ClueInput component
   const submitClue = (clueToShare, clueNumberToShare) => {
     // emit clue
-    shareClue(clueToShare, clueNumberToShare);
+    socket.to(room).emit('clue', {clue: clueToShare, clueNumber: clueNumberToShare});
     // setEditing to false
     setEditing(!editing);
-    // setActiveTeam to the other team
-    passTurn();
   }
-
-  useEffect(() => {
-    // listen for latest clue...
-    // be sure to clean up
-  }, []);
+  socket.on('clue', data => {
+    setClue(data.clue);
+    setClueNumber(data.clueNumber);
+  })
 
   { (!editor && !editing)
     ?
